requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
protobuf==3.19.1
python-dotenv==0.19.1
Werkzeug==2.0.2
